first (or third) commit
this is technically my third commit, but due to a spelling error in the git file i had to remake the repository. this is a Classic RPG style game project. i've added a inventory system with items you can pick up, move in the inventory,  save it and even destroy items. i've also added a character that can walk around and pick items up.